+++ kgrid/cli.py
@@ -46,6 +46,13 @@ def calc_grid(cutoff_length,
 
 def main():
     parser = ArgumentParser()
+    parser.add_argument(
+        action="store",
+        nargs="?",
+        type=str,
+        dest="file",
+        default="geometry.in",
+        help="Path to input file [default: ./geometry.in]")
     threshold = parser.add_mutually_exclusive_group()
     threshold.add_argument(
         "-c",
@@ -68,16 +75,16 @@ def main():
         action="store",
         type=float,
         dest="kspacing",
-        help="Reciprocal space distance like KSPACING in VASP")
-
-    parser.add_argument(
-        "-f",
-        "--file",
+        help="Reciprocal-space distance like KSPACING in VASP")
+    threshold.add_argument(
+        "--castep",
+        "--castep_spacing",
+        "--castep_mp_spacing",
         action="store",
-        type=str,
-        dest="file",
-        default="geometry.in",
-        help="Path to input file [default: ./geometry.in]")
+        type=float,
+        dest="castep_mp_spacing",
+        help=("Reciprocal-space distance like KPOINTS_MP_SPACING in CASTEP; "
+              "this differs from Vasp-like KSPACING by factor of 1/(2 pi)."))
     parser.add_argument(
         "-t",
         "--type",
@@ -91,7 +98,7 @@ def main():
         action="store_true",
         help="Use real-space vector lengths instead of "
         "computing reciprocal cell; not recommended!")
-    # Add further options here
+
     args = parser.parse_args()
 
     if args.vasp_auto:
@@ -100,6 +107,9 @@ def main():
     elif args.kspacing:
         mode = 'kspacing'
         cutoff = args.kspacing
+    elif args.castep_mp_spacing:
+        mode = 'castep_mp_spacing'
+        cutoff = args.castep_mp_spacing
     else:
         mode = 'default'
         cutoff = args.cutoff_length